Weekly Devlog #1
Started working on my competition project for TSA and FBLA, basically just an application that allows your local CTSO chapters to organize and categorize data on local business partners that they are working with or are going to work with, as well as having things like a calendar to have meetings with these business partners. I have to finish this in literally like a week so we will see how it goes. Oh yeah and I also started using emacs for programming because I find it better than most other IDEs(aside from compilation).